+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/Orange-OpenSource/ouds-ios/compare/0.9.0...develop)
 
+### Added
+
+- [Library] Link component tokens ([#384](https://github.com/Orange-OpenSource/ouds-ios/issues/384))
+
 ## [0.9.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.8.0...0.9.0) - 2025-01-09
 
 ### Added